+++ forui/CHANGELOG.md
@@ -25,6 +25,8 @@
 * Change `FTappableStyle.mouseCursor` to `MouseCursor.defer`. See https://ux.stackexchange.com/questions/105024/why-dont-button-html-elements-have-a-css-cursor-pointer-by-default
   for our rationale behind this change.
 
+* Fix `FTabs` throwing an assertion error if `FTabController` is provided with a `initialIndex` > 0.
+
 
 ### 0.14.1
 * Fix `FToaster`sometimes crashing due to an incorrect update of a late final variable.